Refactor base test to separate between tests with login and without login
