+++ SeleniumTests/BaseTests/BaseTest.cs
@@ -1,8 +1,6 @@
 ﻿using NUnit.Framework;
 using NUnit.Framework.Interfaces;
 using SeleniumFramework;
-using SeleniumFramework.Pages.PiguLT;
-using SeleniumFramework.Pages.PiguLTPage;
 
 namespace SeleniumTests.BaseTests
 {
@@ -14,16 +12,6 @@ namespace SeleniumTests.BaseTests
             Driver.InitializeDriver();
             Driver.OpenPage("https://pigu.lt/lt/");
             Driver.MaximizeWindow();
-
-            // Login to page
-            // If you get the error "Too many logins, please try later." when trying to login, change the "EnterLoginEmail" variable. Select one from four:
-            // InputEmail1; InputEmail2; InputEmail3; InputEmail4;
-            HomePage.ClickAllowAllCookies();
-            HomePage.HoverOnProfileIcon();
-            HomePage.ClickOnBlueLoginButton();
-            LoginPage.EnterLoginEmail(TestData.Login.InputEmail3);
-            LoginPage.EnterLoginPassword(TestData.Login.InputPassword);
-            LoginPage.ClickOnGreenLoginButton();
         }
 
         [TearDown]
